Move 'utility' to ze_util.{h,c}
